pantheon-base/panthon-files: fix hard-coded "lib"
The pantheon-files ebuilds install various dynamically-loaded .so
files into /usr/lib due to these locations being hard-coded in the
CMakeFiles.  This fixes this by hacking up the CMakeFiles to point
to $(get_libdir) instead with a sed script.

TODO: presumably some pantheon source-code (perhaps in a different
package?) also looks to these hard-coded /usr/lib paths.  Annoyingly,
however, this probably won't result in a nice clean failure.  In
practice, /usr/lib is a symlink to $(libdir) for everyone for whom
they're not one and the same inode.  If so this should be found and
resolved.

Signed-off-by: Greg Turner <gmt@be-evil.net>