Continue to remove arm KEYWORD.
No reason to keep arm since we have no one actively testing it. We'll
add it back in later.

Repoman is continuing to complain about various packages being in
RDEPEND. I have dealt with some, but I need to do some testing with
which packages on a clean system (probably in a chroot) to verify. I'm
fairly certian there's a lot of packages in RDEPEND that need to be
slotted, or moved to DEPEND, or (probably best) to put them in CDEPEND.